Work on proper synth functionality

+++ wasm_backend/src/exports.zig
@@ -16,7 +16,7 @@ export fn deinitAudio() void {
     audio_processor_web.deinit();
 }
 
-export fn prepareAudio(sample_rate: f64, num_channels: u32, block_size: u32) bool {
+export fn prepareAudio(sample_rate: f64, num_channels: usize, block_size: usize) bool {
     return audio_processor_web.prepare(.{
         .sample_rate = sample_rate,
         .num_channels = num_channels,
